+++ tests/api_tests.h
@@ -36,9 +36,9 @@
 typedef union {
     IO_STRUCT
     {
-        int64_t security;
-        int64_t bip32_path_length;
-        int64_t bip32_path[BIP32_PATH_LENGTH];
+        uint8_t security;
+        uint8_t bip32_path_length;
+        uint32_t bip32_path[BIP32_PATH_LENGTH];
     }
     a;
     SET_SEED_INPUT b;
